+++ library/CHANGELOG.md
@@ -2,6 +2,10 @@
 
 All notable changes to the library will be documented in this file.
 
+## vX.X.X (Month DD, YYYY)
+
+- Rename `action` argument and property of `transform` and `transformAsync` action to `operation`
+
 ## v0.33.3 (June 19, 2024)
 
 - Fix `_isPartiallyTyped` util of `partialCheck` and `partialCheckAsync` action